requirements.txt
langchain-groq
python-dotenv
gradio
pypdf
requests
groq